requirements.txt
fastapi==0.95.2
starlette==0.27.0
uvicorn
pydantic==1.10.14
typing-extensions==4.7.1
anyio==3.7.1
sniffio==1.3.0
aiohttp
attrs
yarl
multidict
aiosignal
async-timeout
charset-normalizer
requests
h11
beautifulsoup4
lxml
ujson
colorlog
nest_asyncio
schedule
pycryptodome
yt-dlp
pytubefix
ffmpeg-python
websockets
aiogram>=3.0.0
aiogram-i18n
psutil
humanize
speedtest-cli
GitPython
